+++ bench/validators/src/stnl.case.ts
@@ -1,28 +1,28 @@
-import { defineCase } from "@utils";
+import { defineCase } from '@utils';
 
-import buildValidateJson from 'stnl/compilers/validate-json/compose';
-import buildStringifyJson from 'stnl/compilers/stringify-json/compose';
+import buildValidateJson from 'stnl/compilers/validate-json/compose.js';
+import buildStringifyJson from 'stnl/compilers/stringify-json/compose.js';
 
 export default defineCase({
   name: 'stnl',
   tests: {
     assertLoose: buildValidateJson({
       props: {
-        number: 'f64',
-        negNumber: 'f64',
-        maxNumber: 'f64',
+        number: 'float',
+        negNumber: 'float',
+        maxNumber: 'float',
         string: 'string',
         longString: 'string',
         boolean: 'bool',
         deeplyNested: {
           props: {
             foo: 'string',
-            num: 'f64',
-            bool: 'bool'
-          }
+            num: 'float',
+            bool: 'bool',
+          },
         },
-        items: { item: 'f64' }
-      }
+        items: { item: 'float' },
+      },
     }),
 
     stringify: buildStringifyJson({
@@ -30,9 +30,9 @@ export default defineCase({
         props: {
           name: 'string',
           pwd: 'string',
-          id: { item: 'f64' }
-        }
-      }
-    })
-  }
+          id: { item: 'float' },
+        },
+      },
+    }),
+  },
 });
